json errors resolved

+++ Advertisement2/src/main/java/com/springdata/restApi/UserRepositories.java
@@ -3,10 +3,10 @@ package com.springdata.restApi;
 	import java.util.List;
 	import org.springframework.data.jpa.repository.JpaRepository;
 	import org.springframework.stereotype.Component;
-	
+
 
 	@Component
-	public interface UserRepositories extends JpaRepository<User, Long> {
+	public interface UserRepositories extends JpaRepository<User, Integer> {
 		
 		List<User> findByFirstName(String firstName);
 		List<User> findByOrderByFirstNameAsc();
